% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bruennert2009_CMLData.R
\name{bruennert2009_CMLData}
\alias{bruennert2009_CMLData}
\title{Obtain the bruennert2009_CML data}
\usage{
bruennert2009_CMLData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the CML bulk RNA-seq data from Bruennert et al. (2009).
}
\details{
Column metadata is provided in the same form as supplied in GSE12211.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/bruennert2009_CML}.
}
\examples{
sce <- bruennert2009_CMLData()

}
\references{
Bruennert D et al. (2009).
Early in vivo changes of the transcriptome in Philadelphia chromosome-positive CD34+ cells from patients with chronic myelogenous leukaemia following imatinib therapy.
\emph{Pharmacogenetics and genomics} 23(5), 983-985
}
\author{
Yi-Wen Hsiao
}
