% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clelland2013_healthyData.R
\name{clelland2013_healthyData}
\alias{clelland2013_healthyData}
\title{Obtain the clelland2013_healthy data}
\usage{
clelland2013_healthyData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the CML bulk RNA-seq data from Clelland et.al. (2013).
}
\details{
Column metadata is provided in the same form as supplied in GSE46449.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/clelland2013_healthy}.
}
\examples{
sce <- clelland2013_healthyData()

}
\references{
Clelland CL et al. (2013).
Utilization of never-medicated bipolar disorder patients towards development and validation of a peripheral biomarker profile.
\emph{PLoS One} 8(6), e69082
}
\author{
Yi-Wen Hsiao
}
