% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/de2010_AMLData.R
\name{de2010_AMLData}
\alias{de2010_AMLData}
\title{Obtain the de2010_AML data}
\usage{
de2010_AMLData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the CML bulk RNA-seq data from de Jonge et.al. (2010).
}
\details{
Column metadata is provided in the same form as supplied in GSE22056.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/de2010_AML}.
}
\examples{
sce <- de2010_AMLData()

}
\references{
de Jonge HJM et al. (2010).
High VEGFC expression is associated with unique gene expression profiles and predicts adverse prognosis in pediatric and adult acute myeloid leukemia.
\emph{Blood, The Journal of the American Society of Hematology} 116(10), 1747-1754
}
\author{
Yi-Wen Hsiao
}
