% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grossmann2012_ALLData.R
\name{grossmann2012_ALLData}
\alias{grossmann2012_ALLData}
\title{Obtain the grossmann2012_ALL data}
\usage{
grossmann2012_ALLData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the CML bulk RNA-seq data from Grossmann et.al. (2012).
}
\details{
Column metadata is provided in the same form as supplied in GSE30599.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/grossmann2012_ALL}.
}
\examples{
sce <- grossmann2012_ALLData()

}
\references{
Grossmann V et al. (2012).
EZH2 mutations and their association with PICALM-MLLT10 positive acute leukaemia.
\emph{British journal of haematology} 157(3), 387-390
}
\author{
Yi-Wen Hsiao
}
