% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/klein2009_AMLData.R
\name{klein2009_AMLData}
\alias{klein2009_AMLData}
\title{Obtain the klein2009_AML data}
\usage{
klein2009_AMLData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the healthy bulk RNA-seq data from Klein et al. (2009).
}
\details{
Column metadata is provided in the same form as supplied in GSE15434.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/klein2009_AML}.
}
\examples{
sce <- klein2009_AMLData()

}
\references{
Klein HU et al. (2009).
A longitudinal study of gene expression in healthy individuals.
\emph{BMC bioinformatics} 1, 1-11
}
\author{
Yi-Wen Hsiao
}
