% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kohlmann2008_allData.R
\name{kohlmann2008_allData}
\alias{kohlmann2008_allData}
\title{Obtain the kohlmann2008_all data}
\usage{
kohlmann2008_allData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the AML, ALL, CML, CLL and healthy bulk RNA-seq data from Kohlmann et al. (2008).
}
\details{
Column metadata is provided in the same form as supplied in GSE13159.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/kohlmann2008_all}.
}
\examples{
sce <- kohlmann2008_allData()

}
\references{
Kohlmann A et al. (2008).
An international standardization programme towards the application of gene expression profiling in routine leukaemia diagnostics: the Microarray Innovations in LEukemia study prephase
\emph{British journal of haematology} 142(5), 802-807
}
\author{
Yi-Wen Hsiao
}
