% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leukemia_refData.R
\name{leukemia_refData}
\alias{leukemia_refData}
\title{Obtain the  leukemia_ref data}
\usage{
leukemia_refData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the combined bulk RNA-seq data from 23 datasets.
}
\details{
Column metadata contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/leukemia_ref}.
}
\examples{
sce <- leukemia_refData()

}
\author{
Yi-Wen Hsiao
}
