% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leukemiaAtlas-package.R
\docType{package}
\name{leukemiasAtlas-package}
\alias{leukemiasAtlas-package}
\alias{leukemiaAtlas-package}
\alias{leukemiaAtlas}
\title{Title}
\description{
This package contains a collection of 38 publicly available bulk or single-cell RNA-seq datasets specific to leukemia disease.
}
\author{
Yi-Wen Hsiao
}
