% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosell2011_healthyData.R
\name{rosell2011_healthyData}
\alias{rosell2011_healthyData}
\title{Obtain the rosell2011_healthy data}
\usage{
rosell2011_healthyData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the healthy bulk RNA-seq data from Rosell et al.(2011).
}
\details{
Column metadata is provided in the same form as supplied in GSE25414.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/rosell2011_healthy}.
}
\examples{
sce <- rosell2011_healthyData()

}
\references{
Ng SW et al. (2022).
Brain perihematoma genomic profile following spontaneous human intracerebral hemorrhage.
\emph{PLoS One} 6(2), e16750
}
\author{
Yi-Wen Hsiao
}
