% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shamir2017_healthyData.R
\name{shamir2017_healthyData}
\alias{shamir2017_healthyData}
\title{Obtain the shamir2017_healthy data}
\usage{
shamir2017_healthyData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the healthy bulk RNA-seq data from Rosell et al.(2017).
}
\details{
Column metadata is provided in the same form as supplied in GSE25414.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/shamir2017_healthy}.
}
\examples{
sce <- shamir2017_healthyData()

}
\references{
Shamir R et al. (2017).
Analysis of blood-based gene expression in idiopathic Parkinson disease.
\emph{Neurology} 89(16), 1676-1683
}
\author{
Yi-Wen Hsiao
}
