% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tasaki2017_healthyData.R,
%   R/tasaki2018_healthyData.R
\name{tasaki2017_healthyData}
\alias{tasaki2017_healthyData}
\title{Obtain the tasaki2017_healthy data}
\usage{
tasaki2017_healthyData()

tasaki2017_healthyData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.

A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the healthy bulk RNA-seq data from Tasaki et al.(2017).

Obtain the healthy bulk RNA-seq data from Tasaki et al.(2018).
}
\details{
Column metadata is provided in the same form as supplied in GSE84844.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/tasaki2017_healthy}.

Column metadata is provided in the same form as supplied in GSE93272.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/tasaki2018_healthy}.
}
\examples{
sce <- tasaki2017_healthyData()

sce <- tasaki2017_healthyData()

}
\references{
Tasaki S et al. (2017).
Multiomic disease signatures converge to cytotoxic CD8 T cells in primary Sjögren’s syndrome.
\emph{Annals of the Rheumatic Diseases} 76(8), 1458-1466

Tasaki S et al. (2018).
Multi-omics monitoring of drug response in rheumatoid arthritis in pursuit of molecular remission.
\emph{Nature communications} 9(1), 2755
}
\author{
Yi-Wen Hsiao
}
