% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warren2009_healthyData.R
\name{warren2009_healthyData}
\alias{warren2009_healthyData}
\title{Obtain the warren2009_healthy data}
\usage{
warren2009_healthyData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the healthy bulk RNA-seq data from Warren et al.(2009).
}
\details{
Column metadata is provided in the same form as supplied in GSE11375.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/warren2009_healthy}.
}
\examples{
sce <- warren2009_healthyData()

}
\references{
Warren HS et al. (2009).
A genomic score prognostic of outcome in trauma patients.
\emph{Molecular medicine} 15, 220-227
}
\author{
Yi-Wen Hsiao
}
