% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zhang2013_CMLData.R
\name{zhang2013_CMLData}
\alias{zhang2013_CMLData}
\title{Obtain the zhang2013_CML data}
\usage{
zhang2013_CMLData()
}
\value{
A \linkS4class{SummarizedExperiment} object with a single matrix of expression values after rank transformation.
}
\description{
Obtain the CML bulk RNA-seq data from Zhang et al.(2013).
}
\details{
Column metadata is provided in the same form as supplied in GSE20876.
This contains information such as sample_id, study_id, platform, subtypes and/or cytogenetics risks.

All data are downloaded from ExperimentHub and cached for local re-use.
Specific resources can be retrieved by searching for \code{leukemiasAtlas/zhang2013_CML}.
}
\examples{
sce <- zhang2013_CMLData()

}
\references{
Zhang B et al. (2013).
Microenvironmental protection of CML stem and progenitor cells from tyrosine kinase inhibitors through N-cadherin and Wnt-β-catenin signaling.
\emph{Blood, The Journal of the American Society of Hematology} 121(10), 1824-1838
}
\author{
Yi-Wen Hsiao
}
